# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Logueo",
        "description": "Sistema de prática con los estudiantes de quinto nivel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/usuario": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Listado de los registro de Usuarios",
                "operationId": "2d3cb90fb7244b81d25f1244015e7648",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "dilan"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "cris@gmail.com"
                                                    },
                                                    "email_varified_at": {
                                                        "type": "timestamp",
                                                        "example": ""
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "132135"
                                                    },
                                                    "estado": {
                                                        "type": "booleano",
                                                        "example": "0"
                                                    },
                                                    "tipo_id": {
                                                        "type": "foreignId",
                                                        "example": "2"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/registrarse": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Registo usuario",
                "operationId": "idcreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "nombre del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Correo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Contrasena",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo_id",
                        "in": "path",
                        "description": "Tipo de usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "402": {
                        "description": "Campos requeridos"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}